ButtonをUIComponentへ移動
